add drawer and fix styles
