+++ AppEntry.js
@@ -1,3 +1,4 @@
+import 'react-native-gesture-handler';
 import { registerRootComponent } from "expo";
 import { Platform } from "react-native";
 import App from "./App";
