+++ package.json
@@ -10,6 +10,8 @@
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.23.1",
+    "@react-navigation/drawer": "^6.6.15",
+    "@react-navigation/native": "^6.1.17",
     "expo": "~50.0.14",
     "expo-status-bar": "~1.11.1",
     "lodash": "^4.17.21",
@@ -18,7 +20,11 @@
     "react": "18.2.0",
     "react-native": "0.73.6",
     "react-native-fetch-api": "^3.0.0",
-    "react-native-polyfill-globals": "^3.1.0"
+    "react-native-polyfill-globals": "^3.1.0",
+    "react-native-safe-area-context": "4.8.2",
+    "react-native-screens": "~3.29.0",
+    "react-native-gesture-handler": "~2.14.0",
+    "react-native-reanimated": "~3.6.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
