+++ tailwind.config.js
@@ -1,9 +1,12 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ["./App.{js,jsx,ts,tsx}", "./<custom directory>/**/*.{js,jsx,ts,tsx}"],
+  content: [
+    './App.{js,jsx,ts,tsx}',
+    './src/Screens/**/*.{js,jsx,ts,tsx}',
+    './src/Components/**/*.{js,jsx,ts,tsx}',
+  ],
   theme: {
     extend: {},
   },
   plugins: [],
-}
-
+};
